refactored decrement rate to int from double

+++ src/main/java/com/subscribers/business/SubscriberService.java
@@ -17,7 +17,7 @@ public class SubscriberService {
 		subscriberDao = new SubscriberDao();
 	}
 
-	public void addSubscriber(PhoneNumber phoneNumber, String name, int balance, double decrementRate)
+	public void addSubscriber(PhoneNumber phoneNumber, String name, int balance, int decrementRate)
 	throws SubscriberAlreadyExistsException, BalanceOutOfRangeException {
 
 		SubscriberBuilder subscriberBuilder = new SubscriberBuilder();
